#217 fixed package names of welcome example
