--- examples/mini-examples/welcome-example/src/main/java/de/saxsys/jfx/exampleapplication/model/Gender.java
@@ -1,13 +1,0 @@
-package de.saxsys.jfx.exampleapplication.model;
-
-/**
- * Enum of possible gender information. If a person doesn't like to provide a gender, the value {@link #NOT_SPECIFIED}
- * can be used.
- */
-public enum Gender {
-	
-	MALE,
-	FEMALE,
-	NOT_SPECIFIED
-	
-}
