--- examples/mini-examples/welcome-example/src/main/java/de/saxsys/jfx/exampleapplication/viewmodel/personlogin/PersonLoginViewModelNotifications.java
@@ -1,19 +1,0 @@
-package de.saxsys.jfx.exampleapplication.viewmodel.personlogin;
-
-public enum PersonLoginViewModelNotifications {
-	OK("Das Einloggen war erfolgreich");
-	
-	private String message;
-	
-	PersonLoginViewModelNotifications(String message) {
-		this.message = message;
-	}
-	
-	public String getMessage() {
-		return message;
-	}
-	
-	public String getId() {
-		return toString();
-	}
-}
