+++ examples/mini-examples/welcome-example/src/main/java/de/saxsys/mvvmfx/examples/welcome/viewmodel/maincontainer/MainContainerViewModel.java
@@ -1,0 +1,22 @@
+package de.saxsys.mvvmfx.examples.welcome.viewmodel.maincontainer;
+
+import javafx.beans.property.ListProperty;
+import javafx.beans.property.SimpleListProperty;
+import javafx.collections.FXCollections;
+
+import de.saxsys.mvvmfx.ViewModel;
+
+public class MainContainerViewModel implements ViewModel {
+	
+	private ListProperty<Integer> displayedPersons = new SimpleListProperty<>(
+			FXCollections.<Integer> observableArrayList());
+	
+	public MainContainerViewModel() {
+		
+	}
+	
+	public ListProperty<Integer> displayedPersonsProperty() {
+		return displayedPersons;
+	}
+	
+}
