--- examples/mini-examples/welcome-example/src/main/resources/de/saxsys/jfx/exampleapplication/view/personlogin/PersonLoginView.fxml
@@ -1,36 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<?import java.lang.*?>
-<?import javafx.collections.*?>
-<?import javafx.scene.control.*?>
-<?import javafx.scene.layout.*?>
-<?import javafx.scene.text.*?>
-
-<StackPane maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity"
-		   xmlns:fx="http://javafx.com/fxml/1" xmlns="http://javafx.com/javafx/8.0.40"
-		   fx:controller="de.saxsys.jfx.exampleapplication.view.personlogin.PersonLoginView">
-	<!-- TODO Add Nodes -->
-	<children>
-		<VBox fx:id="layoutVbox" alignment="CENTER" minHeight="37.0" spacing="5.0">
-			<children>
-				<Label alignment="CENTER" prefWidth="270.0" text="Personen" textAlignment="LEFT">
-					<font>
-						<Font size="30.0"/>
-					</font>
-				</Label>
-				<ChoiceBox id="persons" fx:id="personsChoiceBox" prefWidth="277.0">
-					<items>
-						<FXCollections fx:factory="observableArrayList">
-							<String fx:value="Item 1"/>
-							<String fx:value="Item 2"/>
-							<String fx:value="Item 3"/>
-						</FXCollections>
-					</items>
-				</ChoiceBox>
-				<Button fx:id="loginButton" alignment="CENTER" mnemonicParsing="false" onAction="#loginButtonPressed"
-						text="Login"/>
-			</children>
-		</VBox>
-		<ProgressIndicator fx:id="loginProgressIndicator"/>
-	</children>
-</StackPane>
