+++ examples/mini-examples/welcome-example/src/main/resources/de/saxsys/mvvmfx/examples/welcome/view/maincontainer/MainContainerView.fxml
@@ -1,0 +1,26 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<?import javafx.geometry.*?>
+<?import java.lang.*?>
+<?import javafx.scene.control.*?>
+<?import javafx.scene.layout.*?>
+<?import javafx.scene.control.ListView?>
+<?import javafx.scene.layout.Pane?>
+<?import javafx.scene.layout.VBox?>
+
+<Pane id="personInfoVbox" xmlns:fx="http://javafx.com/fxml/1" xmlns="http://javafx.com/javafx/8.0.40"
+	  fx:controller="de.saxsys.mvvmfx.examples.welcome.view.maincontainer.MainContainerView">
+	<!-- TODO Add Nodes -->
+	<children>
+		<VBox prefHeight="-1.0" spacing="5.0">
+			<children>
+				<!-- Take care for setting the ID to use the injected objects in the code-behind -->
+				<fx:include fx:id="loginView" source="../personlogin/PersonLoginView.fxml"/>
+				<ListView fx:id="personWelcomeListView" blendMode="SRC_OVER" prefHeight="200.0" prefWidth="200.0"/>
+			</children>
+			<padding>
+				<Insets bottom="10.0" left="10.0" right="10.0" top="10.0"/>
+			</padding>
+		</VBox>
+	</children>
+</Pane>
